+++ src/main/java/nextstep/auth/AuthService.java
@@ -2,9 +2,7 @@ package nextstep.auth;
 
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
-import nextstep.support.InvalidMemberException;
-import org.springframework.dao.DataAccessException;
-import org.springframework.dao.EmptyResultDataAccessException;
+import nextstep.support.exception.InvalidMemberException;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -19,14 +17,10 @@ public class AuthService {
 
 
     public String createToken(TokenRequest request) {
-        try{
-            Member member = memberDao.findByUsername(request.getUsername());
-            if(member.checkWrongPassword(request.getPassword())){
-                throw new InvalidMemberException();
-            }
-            return jwtTokenProvider.createToken(member.getId().toString());
-        }catch (EmptyResultDataAccessException exception){
+        Member member = memberDao.findByUsername(request.getUsername());
+        if( member == null || member.checkWrongPassword(request.getPassword()))
             throw new InvalidMemberException();
-        }
+
+        return jwtTokenProvider.createToken(member.getId().toString());
     }
 }
