--- src/main/java/nextstep/support/AuthorizationException.java
@@ -1,11 +1,0 @@
-package nextstep.support;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.UNAUTHORIZED)
-public class AuthorizationException extends RuntimeException {
-    public AuthorizationException() {
-        super("허가되지 않은 접근입니다.");
-    }
-}
