+++ src/main/java/nextstep/support/exception/AuthorizationException.java
@@ -1,0 +1,7 @@
+package nextstep.support.exception;
+
+public class AuthorizationException extends RuntimeException {
+    public AuthorizationException() {
+        super("접근 권한이 없습니다.");
+    }
+}
