+++ src/main/java/nextstep/support/exception/NotExistReservationException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.support.exception;
 
 public class NotExistReservationException extends RuntimeException{
     public NotExistReservationException() {
